xpath: Handle a NodeSet being passed to `id` (#40593)
When a NodeSet is passed to the `id` function then we want to act as if
we were computing the union of individual `id` calls for each node in
the set. We weren't doing that before.

Depends on https://github.com/servo/servo/pull/40592
Testing: A new test starts to pass

---------

Signed-off-by: Simon Wülker <simon.wuelker@arcor.de>