Tweak automatic release notes (#40651)
When drafting github releases, there is an option
to automatically generate a list of changes based on merged PR titles.
For now we simply exclude dependabot PRs, but in the future we might
want to use labels to further optimise the automatic release notes
usefulness.
Documentation:
https://docs.github.com/en/repositories/releasing-projects-on-github/automatically-generated-release-notes#configuring-automatically-generated-release-notes


Testing: Not tested, no impact on servo itself.

---------

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>
Signed-off-by: Jonathan Schwender <55576758+jschwe@users.noreply.github.com>
Co-authored-by: Josh Matthews <josh@joshmatthews.net>