layout: Adjust inset-modified containing block to have non-negative size (#40628)
If the insets of an absolutely positioned box are big enough that the
inset-modified containing block (IMCB) would get a negative size, we
will now adjust the weaker inset so that the size of the IMCB becomes
zero instead.
This affects auto sizes, auto margins, and self-alignment.

This behavior matches Blink for the most part. However, while Blink also
ensures a non-negative IMCB size, it doesn't necessarily do so by
adjusting the weaker inset, so the alignment may differ.

Blink is fine with adopting this behavior, as resolved by the CSSWG in
https://github.com/w3c/csswg-drafts/issues/11478

Testing: Adding new WPT

---------

Signed-off-by: Oriol Brufau <obrufau@igalia.com>