Compositor: Add touch compatibility handling for mouse move events (#40637)
Every time a touchmove event is added, it updates the cursor, causing
frequent hit testing and affecting the performance of scrolling. Adding
a flag to mark the current touchmove is done for compatibility with
touch events. In this case, the last_mouse_move_position will not be
updated.

Testing: No new use cases will be approved; it's just for performance
optimization.
Fixes: A solution for #39433

Signed-off-by: kongbai1996 <1782765876@qq.com>