libservo: Trigger a rendering update when waiting for paint readiness in WebView unit tests (#40712)
Some unit tests attempt to wait for the WebView to be paint ready before
executing. The previous code would wait for a frame ready message from
the WebView before running, but this was subject to timing issues due to
the frame being sent before the load complete message. This changes
makes it so that the WebView explicitly triggers another rendering
update and then waits for a frame. It should be guaranteed that the next
frame will be one that reflects load completeness.

Special thanks to jmunroe on Zulip for investigation into this issue.

Testing: This seems to eliminate flaky timeouts on my machine. I ran
the unit tests in a loop, stopping if any error (such as a timeout
happened).
Before this would stop within a minute or two, but I could not reproduce
the issue with this change after several minutes.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>