ohos CI: Add webdriver based scenario tests (#40473)
Add webdriver based scripts to run scenario tests on OpenHarmony
devices, with the purpose of testing functionality in CI. For now
passing these tests will not be enforced, as we first need to see how
reliable the CI tests are in daily usage.

Testing:
Functional and Performance Test Cases: Test that when new code is
deployed, the Servo homepage and Mossel homepage load correctly;
Mossel's navigation and swipe functions work properly.

---------

Signed-off-by: monissaJ <2362314605@qq.com>
Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>
Co-authored-by: Jonathan Schwender <schwenderjonathan@gmail.com>