Add feature to disable JIT support in mozjs at compile-time (#37972)
When the `js_jit` feature is not selected, this reduces the binary size
from 89 to 80 MiB for linux production-stripped.
By default, we still enable JIT, but in scenarios where JIT is not
required or not allowed by the system, the size savings make a feature
worthwhile.
To allow disabling the feature, `servoshell` now depends on libservo
with `no-default-features`, and hence gains explicit default features,
that were previously enabled via libservos default features. To avoid
compile-errors with `--no-default-features`, `vello-cpu` is
unconditionally selected by servoshell (identical to previous behavior).

Testing: The default behavior is unchanged. The HarmonyOS CI job does
test with `--no-default-features`

---------

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>