fonts: Compute applied variations depending on originating `@font-face` rule (#40647)
This change gives the font selection code access to the data from the
`@font-face` rule that defined the font - and uses it to apply
`font-variation-settings` inside `@font-face` rules.

Testing: We don't run wpt with `layout_variable_fonts_enabled=true` yet.
I've attached a test case that is fixed by this PR.
Part of #37236

<details><summary>Testcase</summary>

```html
<!DOCTYPE html>
<html lang="en"><head>
    <style>
        h1, h2, h3, h4, h5, h6 {
            font-family: 'Montserrat', sans-serif;
        }
    </style>
    <style>
        @font-face {
          font-family: 'Montserrat';
          font-display: swap;
          font-variation-settings: "wght" 500;
          src: url(https://fonts.gstatic.com/s/montserrat/v30/JTUSjIg1_i6t8kCHKm459Wlhyw.woff2) format('woff2');
        }
    </style>
    <h2>The hackfest</h2>
</body></html>
```
</details>

---------

Signed-off-by: Simon Wülker <simon.wuelker@arcor.de>