libservo: Remove `ServoDelegate::notify_animating_changed` (#40886)
This is dead code #40799, as it is never called. Remove the delegate
call and the entire `ServoShellServoDelegate` as the only thing that it
was doing was printing top-level errors which only happen when something
goes very wrong. It's quite likely that backtraces printed earlier
will be more useful.

Testing: This just removes dead code, so should be covered by existing
tests.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>