servoshell: Fix stdout / stderr output in windows production builds (#40961)
When servo is started from the command line, we still want output to be
printed. On windows, due to us using the `windows` subsystem (as opposed
to the default `console` subsystem), this doesn't work out-of-the-box,
and we need to manually attempt to attach to the console of the parent
process. If servo was not started from the command line, then the call
will fail, which we can ignore.
Since we can now have the best of both worlds (i.e no popping up console
window, and preserve stdout / stderr when started from a console), we
can now unconditionally use the `windows` subsystem. (On non-windows
platforms, `windows-subsystem` does nothing)

Testing: Manually tested. The original issue does not reproduce when
redirecting stdout / stderr, so I don't see a way to write an automated
test.
Fixes: #40211

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>