script: Make a Layout-safe version of `Element::each_custom_state` (#40743)
`Element::each_custom_state` is called from Layout, but it creates
`DomRef` types on the stack, which can only be done on the main script
thread. Since layout might be doing styling and element consultation on
workers threads, expose a Layout-safe version of this change, which
doesn't have this issue.

Testing: This fixes two panics encountered during WPT tests.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>