servoshell: Prepare November release (0.0.3) (#40962)
As discussed in the tsc-channel, prepare for the next release by cutting
a version on the last day of the month.

Note: The actual release would happen ~ 2 weeks from now, when the blog
post for november is ready.

This commit was produced by running `./mach release 0.0.3`.

Testing: Not required. Additional pre-release testing will happen after
the version bump and before the release, with potential patches being
cherry-picked to a release branch.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>