Implement usage of custom protocol handlers (#40803)
This implements a dummy `mailto:` protocol handler in
Servoshell. It registers a new custom handler that
fetches content by substituting the registered url
for the protocol handler.

This does not pass any WPT tests, as that requires
using the `document.protocol_handler_automation_mode`
which will be part of a follow-up PR.

Part of #40615

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>