fonts: Replace deprecated `xml-rs` with `xml@1.1` (#41134)
This PR updates `xml-rs` crate. On the official crate site it is
mentioned to change the name of the crate with `xml` and update it.It is
also explicitly mentioned that their won't be any code changes so I
think their isn't any major API changes.

Reference:- https://crates.io/crates/xml-rs/0.8.28

Testing: It doesn't require testing because this is just the change in
the crate's name

---------

Signed-off-by: Ajay Singh <ajaykripa8736968359@gmail.com>
Signed-off-by: Ajay Singh <37625112+Ajay-singh1@users.noreply.github.com>