layout: Merge BaseFragment and BaseFragmentInfo (#41177)
These structs used to be different when they were added in #29699:
`BaseFragment` had a `debug_id` field, while `BaseFragmentInfo` didn't.
But this field was later removed in #35001, so the structs became
identical.

Therefore, this patch just unifies them. I'm picking `BaseFragmentInfo`
as the name, since it was the most frequently used of the two.

Testing: Not needed, no change in behavior

Signed-off-by: Oriol Brufau <obrufau@igalia.com>