layout: Preserve cached intrinsic inline sizes in more cases (#41160)
If a box has layout damage, but the intrinsic contribution of some
ancestor doesn't depend on its contents, then we don't need to clear the
cached intrinsic sizes of further ancestors.

Testing: No behavior change, just an optimization. It would be good to
have unit tests that check we don't do unnecessary work, but leaving
this for the future.

Signed-off-by: Oriol Brufau <obrufau@igalia.com>
Co-authored-by: Martin Robinson <mrobinson@igalia.com>