ohos: Warn on common commandline argument mistakes (#41203)
Currently, commandline arguments need to be used like this:
'--psn=--tracing_filter' which people can easily forget and use
'--psn=tracing_filter'.
This happened to multiple people and leads to weird issues, since the
argument is silently dropped.
Because OHOS appends a lot of other arguments and we don't have a good
way to filter out the OS provided arguments, we now print a big error
message that hopefully should be noticed by the user for a couple of
commonly used arguments.

Testing: Tested on device with and without the correct way to invoke
arguments.

Signed-off-by: Narfinger <Narfinger@users.noreply.github.com>

---------

Signed-off-by: Narfinger <Narfinger@users.noreply.github.com>