net: Make parsing of Access-Control-Allow-Origin header less strict (#41210)
The typed header parsing is too strict. It parses the value as an
origin, which is too strict. Instead, the spec expects a literal string
comparison, regardless if the value reflects a valid origin or not.

Initially I thought this was a spec issue
(https://github.com/whatwg/fetch/issues/1895), but alas it wasn't.
Testing: This allows multiple WPT tests to pass.

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>