libservo: Allow responding to auxiliary `WebView` open requests asynchronously (#41318)
This change renames the `WebView::request_open_auxiliary_webview` to
`WebView::request_create_new` and allows responding to it
asynchronously. A new `CreateNewWebViewRequest` is exposed that keeps
the response sender alive as long as it is not dropped. If dropped
without a response, the original request is denied with `None`.

This also makes the API a bit harder to misuse, because the public
`WebViewBuilder::new_auxiliary` factory is no longer exposed, in favor
of `CreateNewWebViewRequest::builder`. This makes it impossible to
create an auxiliary `WebView` without a corresponding request. This
still isn't the idea API, which would be to make creating an auxiliary
`WebView` and normal `WebView` exactly the same, but it's an
improvement.

This will allow opening auxiliary `WebView`s in new toplevel platform
windows in servoshell. That will be handled in a followup. In addition,
the `CreateNewWebViewRequest` can eventually hold the arguments passed
to `window.open`.

Testing: Corresponding changes are made in servoshell, which means that
this is tested by the WPT calling `window.open`.
Fixes: #41220.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>