CI: Fix ohos speedometer failing (#41349)
Speedometer is failing on main after recent changes. The symptom is that
when creating the webdriver (in python on the host) servo switches to
`about:blank`. This is not expected by the script and hence we fail to
find the start button, since the expected page is the speedometer page.
The quick fix is just navigating to the page after creating the driver,
but we also take this opportunity to slightly refactor this script and
use the common method for running tests, which allows removing some
code.

Testing: Tested locally and via [mach
try](https://github.com/servo/servo/actions/runs/20303863788)
Fixes: #41348

---------

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>