Fix build of `./mach test-unit -p base` (#41498)
`malloc_size_of` depends on `tendril` having the `encoding_rs` feature
turned on, but this was only turned on transitively via `html5ever` and
`script`/ `script_bindings` `tendril` dependencies. This change makes it
so that `malloc_size_of` depends directly on `tendril` and moves this
dependency to the root `Cargo.toml`.

Testing: This just fixes a particular build configuration.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>