script: Introduce concept of a fetch group (#41423)
We now keep track of deferred fetches in globalscope
rather than in document loader. This matches what
the spec expects with its concept of a fetchgroup.

Cancellations are still present in a document loader
and don't match yet all places where we perform fetches.
Therefore, these will leave in document loader for now
until we rearranged the fetch record tracking in the
fetch group (which is a separate field from deferred
fetches).

Fixes #41286

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>